Add storage stage and bank_forks tests to integration
